# src/main/resources/static/api-docs/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/admin/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Recupera um admin pelo ID",
        "operationId": "buscarAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          },
          "404": {
            "description": "Admin não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Atualiza um admin",
        "operationId": "atualizarAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Admin inválido"
          },
          "200": {
            "description": "Admin Atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Deleta um admin",
        "operationId": "deletarAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin deletado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          },
          "404": {
            "description": "Admin não encontrado"
          }
        }
      }
    },
...
    "/v1/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Recupera todos os admins",
        "operationId": "buscarTodosAdmins",
        "responses": {
          "200": {
            "description": "Admin encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          },
          "404": {
            "description": "Admin não encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Cria um admin",
        "operationId": "criarAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Admin inválido"
          },
          "201": {
            "description": "Admin Criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/email/{email}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Atualiza um admin",
        "operationId": "buscarAdminPeloEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Admin inválido"
          },
          "200": {
            "description": "Admin Atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminDTO": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}